fix(ci): clear two rust-1.96 clippy-drift lints (-D warnings)
Pre-existing latent lints unmasked once check-cpp + the example matrix were
reached (the earlier serde_core/cpp failures stopped CI before these steps):

- nros-cpp/src/lib.rs:570 `unnecessary_cast` — `NROS_CPP_RET_INVALID_ARGUMENT
  as i32` on an already-`c_int` const; drop the cast.
- nros::node! `__nros_component_<pkg>_install` `not_unsafe_ptr_arg_deref` — the
  C-ABI export takes mandatory raw-ptr params + an `unsafe` body; add a scoped
  `#[allow(clippy::not_unsafe_ptr_arg_deref)]` (covers every node! pkg).

`just check` (full clippy gate: fast + build + embedded + per-feature +
examples + cpp) green.

+++ packages/core/nros-cpp/src/lib.rs
@@ -567,7 +567,7 @@ pub unsafe extern "C" fn nros_board_native_run_components(
 ) -> i32 {
     let setup = match setup {
         Some(f) => f,
-        None => return NROS_CPP_RET_INVALID_ARGUMENT as i32,
+        None => return NROS_CPP_RET_INVALID_ARGUMENT,
     };
 
     // Env overlay (mirrors the C++ `nros::init()` hosted fallback).
